requirements.txt
langchain_google_genai
crewai
load_dotenv
crewai_tools
